add unit and integration tests for User, Author, Book functionality + required refactoring of related classes and methods

+++ bookShopApp/src/main/java/com/example/bookshopapp/data/book/Book.java
@@ -36,7 +36,7 @@ public class Book {
     private LocalDate pubDate;
 
     @Column(name = "is_bestseller", columnDefinition = "smallint")
-    private int isBestseller;
+    private Float isBestseller;
     private String slug;
     private String title;
     private String image;
